+++ db/events.php
@@ -6,6 +6,10 @@ $observers = array (
         'callback' => 'plagiarism_plagscan_observer::assignsubmission_file_uploaded'
     ),
     array(
+        'eventname' => '\mod_assign\event\assessable_submitted',
+        'callback' => 'plagiarism_plagscan_observer::mod_assign_assessable_submitted'
+    ),
+    array(
         'eventname' => '\assignsubmission_onlinetext\event\assessable_uploaded',
         'callback' => 'plagiarism_plagscan_observer::assignsubmission_onlinetext_uploaded'
     )